% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_genes_by_pathway.R
\name{check_pathway_name}
\alias{check_pathway_name}
\title{Check whether pathway names are valid}
\usage{
check_pathway_name(pathway, class = c("mmu", "hsa", "gga"))
}
\arguments{
\item{pathway}{A vector contains interested pathwat names}

\item{class}{Specify which type of gene}
}
\value{
A vector of path IDs
}
\description{
Check whether pathway names are valid
}
\examples{
Path IDs for each pathway
pathway <- c("Purine metabolism", "PI3K-Akt signaling pathway", "AMPK signaling pathway", "Choline metabolism in cancer")
path.ids = check_pathway_name(pathway, class = "mmu")
}
