% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_deg_from_cuffdiff.R
\name{grab_degs_from_cuffdiff}
\alias{grab_degs_from_cuffdiff}
\title{Grab differential expressed genes (DEGs) from cuffdiff result}
\usage{
grab_degs_from_cuffdiff(
  gene_exp.diff,
  out.name = NULL,
  class = c("mmu", "gga", "hsa"),
  criterion = c("p_value", "q_value"),
  cut.off = 0.05
)
}
\arguments{
\item{gene_exp.diff}{a dataframe from cuffdiff result}

\item{out.name}{If specified, the output will be written to this file.}

\item{criterion}{p_value or q_value}

\item{cut.off}{significance level}
}
\value{
A dataframe contains all DEGs with gene symvbols.
}
\description{
Grab differential expressed genes (DEGs) from cuffdiff result
}
\examples{
head(gene_exp.diff)
deg = grab_deg_from_cuffdiff(gene_exp.diff)
head(deg)
}
