% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeometric_test.R
\name{pathway_analysis}
\alias{pathway_analysis}
\title{hypergeometric test}
\usage{
pathway_analysis(sig.genes, class = c("mmu", "hsa", "gga"))
}
\arguments{
\item{sig.genes}{A vector of significantly changed genes}

\item{class}{Specify which type of gene}
}
\value{
A list of all significantly changed pathways, and all pathways with p value
}
\description{
hypergeometric test
}
\examples{
pathway_analysis(c("Osr1","Tbx5"),"mmu")
}
