% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeometric_test.R
\name{pathway_analysis_meth}
\alias{pathway_analysis_meth}
\title{hypergeometric test for DNA-methylation seq}
\usage{
pathway_analysis_meth(
  sig.genes,
  total.genes,
  class = c("mmu", "hsa", "gga"),
  mc.cores = 8
)
}
\arguments{
\item{sig.genes}{A vector of significantly changed genes}

\item{total.genes}{A vector of total genes}

\item{class}{Specify which type of gene}

\item{mc.cores}{Specify how many cores to use}
}
\value{
A list of all significantly changed pathways, and all pathways with p value
}
\description{
hypergeometric test for DNA-methylation seq
}
\examples{
path_res = pathway_analysis_meth(sig.genes, class = "mmu")
}
